Moved code to md code to central location.
