+++ src/lib/create_post_schema.ts
@@ -3,7 +3,7 @@ import { z } from "zod";
 const createPostSchema = z.object({
   title: z.string().nonempty().min(2).max(50),
   author: z.string().nonempty().min(2).max(50),
-  content: z.string().nonempty().min(10).max(2000),
+  content: z.string().nonempty().min(10).max(10000),
 });
 
 export default createPostSchema;
