+++ src/lib/markdown/blog-post.css
@@ -17,6 +17,22 @@ h2:not(.not_md) {
   @apply text-2xl;
 }
 
+h3:not(.not_md) {
+  @apply text-xl;
+}
+
+h4:not(.not_md) {
+  @apply text-lg;
+}
+
+h5:not(.not_md) {
+  @apply text-base;
+}
+
+h6:not(.not_md) {
+  @apply text-sm;
+}
+
 pre:not(.not_md) {
-  @apply rounded-md bg-black/50 p-4;
+  @apply mb-4 mt-4 rounded-md border border-gray-700 bg-gray-900 p-4 text-sm;
 }
